test(ai): retire historical conformance locks
